#11 - Refactoring and tests parsing - class FolderUtils renamed to RepositoryUtils; - getFilePathForComplexOvalObjects moved to RepositoryUtils; - added parseTestDependencies method for finding object IDs and state IDs from tests.
